shanoir-issue#1174 Remove finish import step
- Dicom import
- EEG import
- Bruker import (+simplify with dicom)
- processed dataset
