+++ shanoir-ng-front/src/app/import/clinical-context/clinical-context.component.html
@@ -226,4 +226,4 @@ along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
     </ol>
 
 </fieldset>
-<button class="next" [disabled]="!valid" (click)="next()">Next</button>
+<button class="next" [disabled]="!valid" (click)="next()">Import now</button>
