+++ shanoir-ng-front/src/app/import/eeg-clinical-context/eeg-clinical-context.component.html
@@ -148,4 +148,4 @@ along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
     [getPage]="getPage.bind(this)" 
     [columnDefs]="columnDefs">
 </shanoir-table>
-<button class="next" [disabled]="!valid" (click)="next()">Next</button>
+<button class="next" [disabled]="!valid" (click)="next()">Import now</button>
