+++ shanoir-ng-front/src/app/import/eeg-clinical-context/eeg-clinical-context.component.ts
@@ -39,6 +39,11 @@ import { ContextData, ImportDataService } from '../shared/import.data-service';
 import { Option } from '../../shared/select/select.component';
 import { AcquisitionEquipmentPipe } from '../../acquisition-equipments/shared/acquisition-equipment.pipe';
 import { Subscription } from 'rxjs';
+import { EegImportJob } from '../shared/eeg-data.model';
+import { MsgBoxService } from '../../shared/msg-box/msg-box.service';
+import { SubjectService } from '../../subjects/shared/subject.service';
+import { ImportService } from '../shared/import.service';
+import { EegDatasetDTO } from '../../datasets/shared/dataset.dto';
 
 
 @Component({
@@ -88,7 +93,10 @@ export class EegClinicalContextComponent implements OnInit {
             private breadcrumbsService: BreadcrumbsService,
             private importDataService: ImportDataService,
             public acqEqPipe: AcquisitionEquipmentPipe,
-            public subjectExaminationPipe: SubjectExaminationPipe) {
+            public subjectExaminationPipe: SubjectExaminationPipe,
+            private subjectService: SubjectService,
+            private importService: ImportService,
+            private msgService: MsgBoxService) {
 
         this.coordSystemOptions = CoordSystems.options;
 
@@ -407,7 +415,56 @@ export class EegClinicalContextComponent implements OnInit {
     }
 
     public next() {
-        this.router.navigate(['imports/eegfinish']);
+        this.startEegImportJob();
+    }
+    
+    public startEegImportJob(): void {
+        let context = this.importDataService.contextData;
+        this.subjectService
+            .updateSubjectStudyValues(context.subject.subjectStudy)
+            .then(() => {
+                let that = this;
+                this.importData()
+                    .then((importJob: EegImportJob) => {
+                        this.importDataService.reset();
+                        setTimeout(function () {
+                            that.msgService.log('info', 'The import successfully started')
+                        }, 0);
+                        // go back to the first step of import
+                        this.router.navigate(['/imports/eeg']);
+                    }).catch(error => {
+                        throw error;
+                    });
+            }).catch(error => {
+                throw new Error('Could not save the subjectStudy object, the import job has been stopped. Cause : ' + error);
+            });
+    }
+
+    private importData(): Promise<any> {
+        let importJob = new EegImportJob();
+        importJob.datasets = [];
+        let context = this.importDataService.contextData;
+        let importJobContext = this.importDataService.eegImportJob;
+
+        for (let dataset of importJobContext.datasets) {
+            let datasetToSet = new EegDatasetDTO();
+            datasetToSet.channels = dataset.channels;
+            datasetToSet.name = dataset.name;
+            datasetToSet.files = dataset.files;
+            datasetToSet.events = dataset.events;
+            datasetToSet.samplingFrequency = dataset.samplingFrequency;
+            datasetToSet.channelCount = dataset.channelCount;
+            datasetToSet.coordinatesSystem = context.coordinatesSystem;
+            importJob.datasets.push(datasetToSet);
+        }
+        importJob.subjectId = context.subject.id;
+        importJob.subjectName = context.subject.name;
+        importJob.studyName = context.study.name;
+        importJob.workFolder = importJobContext.workFolder;
+        importJob.examinationId = context.examination.id;
+        importJob.studyId = context.study.id;
+        importJob.acquisitionEquipmentId = context.acquisitionEquipment.id;
+        return this.importService.startEegImportJob(importJob);
     }
     
     ngOnDestroy() {
