--- shanoir-ng-front/src/app/import/eeg-finish/eeg-finish.component.css
@@ -1,3 +1,0 @@
-.loader { vertical-align: middle; height: 12px; }
-button { margin: 10px 0; }
-img { margin-left: 10px; }
