+++ shanoir-ng-front/src/app/import/processed-dataset-clinical-context/processed-dataset-clinical-context.component.html
@@ -103,4 +103,4 @@ along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
     </ol>
 
 </fieldset>
-<button class="next" [disabled]="!valid" (click)="next()">Next</button>
+<button class="next" [disabled]="!valid" (click)="next()">Import now</button>
