--- shanoir-ng-front/src/app/preclinical/importBruker/finish/bruker-finish.component.css
@@ -1,3 +1,0 @@
-.loader { vertical-align: middle; height: 12px; }
-button { margin: 10px 15px; }
-img { margin-left: 10px; }
