--- shanoir-ng-front/src/app/preclinical/importBruker/finish/bruker-finish.component.html
@@ -1,17 +1,0 @@
-<!--
-Shanoir NG - Import, manage and share neuroimaging data
-Copyright (C) 2009-2019 Inria - https://www.inria.fr/
-Contact us on https://project.inria.fr/shanoir/
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-You should have received a copy of the GNU General Public License
-along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
--->
-
-<div class="header command-zone">4. Import the data</div>
-<button [disabled]="importing" (click)="startImportJob()">Import now</button>
-<span class="progress" *ngIf="importing">Importing in progress ...<img class="loader" [src]="ImagesUrlUtil.SMALL_LOADER_IMAGE_PATH"/></span>
