+++ shanoir-ng-front/src/app/preclinical/preclinical-routing.module.ts
@@ -36,7 +36,6 @@ import { AuthAdminOrExpertGuard } from '../shared/roles/auth-admin-or-expert-gua
 
 import { BrukerUploadComponent } from './importBruker/bruker-upload/bruker-upload.component';
 import { BrukerSelectSeriesComponent } from './importBruker/select-series/bruker-select-series.component';
-import { BrukerFinishImportComponent } from './importBruker/finish/bruker-finish.component';
 import { SubjectTherapyFormComponent } from './therapies/subjectTherapy/edit/subjectTherapy-form.component';
 import { SubjectTherapiesListComponent } from './therapies/subjectTherapy/list/subjectTherapy-list.component';
 import { AnestheticIngredientFormComponent } from './anesthetics/ingredients/edit/anestheticIngredient-form.component';
@@ -65,9 +64,6 @@ let routes : Routes = [
             }, {
                 path: 'series',
                 component: BrukerSelectSeriesComponent
-            }, {
-                path: 'finish',
-                component: BrukerFinishImportComponent
             }
         ]
     },
